+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.0.1 (pending)
 
+* if ronn is not installed locally: invoke ronn via Docker when available
 * update lazy-foss-parent 1.0.1 to 1.3.1
 * update sushi 3.1.2 to 3.2.2